backend/requirements.txt
Django~=4.2.3
djangorestframework~=3.14.0
Pillow~=10.0.0
drf-yasg~=1.21.6
django-cors-headers~=4.2.0
psycopg2~=2.9.6
